Refinery (refinery v0.1.0)

Documentation for Refinery.

Functions

build(module, type, refinements \\ %{})

Takes a schema type and returns an instance of it with the following layers of data applied:

Defaults for this type defined as refinement(type, :default)

refinements passed as second argument

For all referenced associations, a) Defaults for the association type defined in build/1 b) Defaults passed by the previous layer, e.g. %Offer{department: %{name: "Construction"}} will override the :name on the associated Department c) refinements passed for this specific association name, e.g. passing %{department: %{name: "Carpentry"}} will override the :name of any Department directly or indirectly associated in the current branch

create(module, type, refinements \\ %{})

do_build(module, type, attrs)